Make desched ioctls no-ops during replay
During replay all syscalls cause traps to rr (2 per syscall). Each may-block
syscall-buffered syscall is enclosed in an arm/disarm ioctl pair, so that's
4 extra traps per application syscall. Fix this by making all modifying the
rr page during replay so that privileged untraced syscalls just clear the
syscall result register instead of actually performing a syscall. (Currently
only the arm/disarm ioctls are privileged and untraced.) Then, during replay
these ioctls do not enter the kernel and do not cause traps to rr.

We need to adjust the replay code to not advance to these syscalls during
replay.

+++ src/preload/preload.c
@@ -872,16 +872,16 @@ static void arm_desched_event(void) {
    * avoid! :) Although we don't allocate extra space for these
    * ioctl's, we do record that we called them; the replayer
    * knows how to skip over them. */
-  if (privileged_untraced_syscall3(SYS_ioctl, desched_counter_fd,
-                                   PERF_EVENT_IOC_ENABLE, 0)) {
+  if ((int)privileged_untraced_syscall3(SYS_ioctl, desched_counter_fd,
+                                        PERF_EVENT_IOC_ENABLE, 0)) {
     fatal("Failed to ENABLE counter %d", desched_counter_fd);
   }
 }
 
 static void disarm_desched_event(void) {
   /* See above. */
-  if (privileged_untraced_syscall3(SYS_ioctl, desched_counter_fd,
-                                   PERF_EVENT_IOC_DISABLE, 0)) {
+  if ((int)privileged_untraced_syscall3(SYS_ioctl, desched_counter_fd,
+                                        PERF_EVENT_IOC_DISABLE, 0)) {
     fatal("Failed to DISABLE counter %d", desched_counter_fd);
   }
 }
